requirements.txt
crawl4ai==0.4.248
nest_asyncio
requests
supabase>=2.13.0
playwright
google-generativeai
duckduckgo-search
python-dotenv>=1.0.0
pytest
openai
PyYAML
httpx>=0.25.0
python-dotenv>=1.0.0